refactor: extract state logic into a hook
