+++ src/pages/teachers/programmes/[programmeSlug]/units.tsx
@@ -1,5 +1,4 @@
 import React, { useId, useState, useRef, useEffect } from "react";
-import { useRouter } from "next/router";
 import {
   GetStaticPathsResult,
   GetStaticProps,
@@ -18,7 +17,6 @@ import {
 import { examboards, tierSlugs } from "@oaknational/oak-curriculum-schema";
 import { z } from "zod";
 
-import { FilterTypeValueType } from "@/browser-lib/avo/Avo";
 import {
   getFallbackBlockingConfig,
   shouldSkipInitialBuild,
@@ -53,17 +51,18 @@ import CurriculumDownloadBanner from "@/components/TeacherComponents/CurriculumD
 import { convertSubjectToSlug } from "@/components/TeacherComponents/helpers/convertSubjectToSlug";
 import { getMvRefreshTime } from "@/pages-helpers/curriculum/docx/getMvRefreshTime";
 import { isUnitListData } from "@/components/TeacherComponents/UnitList/helpers";
+import { useUnitFilterState } from "@/hooks/useUnitFilterState";
 
 export type UnitListingPageProps = {
   curriculumData: UnitListingData;
   curriculumRefreshTime: number;
 };
 
-export type FilterQuery = {
-  year?: string | null;
-  category?: string | null;
-  theme?: string | null;
-};
+// export type FilterQuery = {
+//   year?: string | null;
+//   category?: string | null;
+//   theme?: string | null;
+// };
 
 const UnitListingPage: NextPage<UnitListingPageProps> = ({
   curriculumData,
@@ -95,10 +94,18 @@ const UnitListingPage: NextPage<UnitListingPageProps> = ({
 
   const learningThemes = curriculumData.learningThemes ?? [];
 
-  const router = useRouter();
-  const themeSlug = router.query["learning-theme"]?.toString();
-  const categorySlug = router.query["category"]?.toString();
-  const yearGroupSlug = router.query["year"]?.toString();
+  const {
+    themeSlug,
+    categorySlug,
+    yearGroupSlug,
+    newFilterQuery,
+    isMobileFilterDrawerOpen,
+    setIsMobileFilterDrawerOpen,
+    handleUpdateActiveFilters,
+    handleUpdateAndSubmitFilterQuery,
+    handleSubmitFilterQuery,
+  } = useUnitFilterState({ isUnitListing: isUnitListData(curriculumData) });
+
   const [skipFiltersButton, setSkipFiltersButton] = useState(false);
   const filtersRef = useRef<HTMLDivElement>(null);
   const isFiltersAvailable =
@@ -180,143 +187,6 @@ const UnitListingPage: NextPage<UnitListingPageProps> = ({
         tierName: tier ?? null,
         examBoard: examBoardTitle,
         pathway: pathwayTitle,
-      });
-    }
-  };
-
-  const [newFilterQuery, setNewFilterQuery] = useState<FilterQuery | null>(
-    null,
-  );
-
-  const handleUpdateActiveFilters = (filters: FilterQuery | null) => {
-    if (!filters) {
-      setNewFilterQuery(null);
-      return null;
-    } else {
-      const params: FilterQuery = Object.assign({}, newFilterQuery);
-      if (filters.year) {
-        params.year = filters.year;
-      } else if (filters.year === null) {
-        params.year = "";
-      }
-      if (filters.category) {
-        params.category = filters.category;
-      } else if (filters.category === null) {
-        delete params.category;
-      }
-      if (filters.theme) {
-        params.theme = filters.theme;
-      }
-      setNewFilterQuery(params);
-      return params;
-    }
-  };
-
-  const trackFilterRefined = (
-    filterType: FilterTypeValueType,
-    filterValue: string,
-  ) => {
-    const activeFilters = {
-      content_types: "units",
-      learning_themes: filterType !== "Learning theme filter" && themeSlug,
-      year_group: filterType !== "Year filter" && yearGroupSlug,
-    };
-    track.browseRefined({
-      platform: "owa",
-      product: "teacher lesson resources",
-      engagementIntent: "refine",
-      componentType: "filter_link",
-      eventVersion: "2.0.0",
-      analyticsUseCase: "Teacher",
-      filterValue,
-      filterType: "Subject filter",
-      activeFilters,
-    });
-  };
-
-  const handleUpdateAndSubmitFilterQuery = (
-    queryObj: FilterQuery | null,
-    filterType: FilterTypeValueType,
-    filterValue: string,
-  ) => {
-    const updatedQuery = handleUpdateActiveFilters(queryObj);
-
-    // TODO: extract to shared
-    const params = Object.assign({}, router.query);
-
-    if (updatedQuery?.year) {
-      params.year = updatedQuery.year;
-    } else {
-      delete params.year;
-    }
-    if (updatedQuery?.category) {
-      params.category = updatedQuery.category;
-    } else {
-      delete params.category;
-    }
-    if (updatedQuery?.theme) {
-      params["learning-theme"] = updatedQuery.theme;
-    } else {
-      params["learning-theme"] = "all";
-    }
-    trackFilterRefined(filterType, filterValue);
-    router.replace(
-      {
-        pathname: router.pathname,
-        query: params,
-      },
-      undefined,
-      { shallow: true },
-    );
-  };
-
-  const [isMobileFilterDrawerOpen, setIsMobileFilterDrawerOpen] =
-    useState<boolean>(false);
-
-  const handleSubmitFilterQuery = () => {
-    if (newFilterQuery) {
-      const params = Object.assign({}, router.query);
-      let theme, category, year;
-      if (newFilterQuery.year) {
-        params.year = newFilterQuery.year;
-        year = newFilterQuery.year;
-      } else {
-        delete params.year;
-      }
-      if (newFilterQuery.category) {
-        params.category = newFilterQuery.category;
-        category = newFilterQuery.category;
-      } else {
-        delete params.category;
-      }
-      if (newFilterQuery.theme) {
-        params["learning-theme"] = newFilterQuery.theme;
-        theme = newFilterQuery.theme;
-      } else {
-        params["learning-theme"] = "all";
-      }
-      if (isUnitListData(curriculumData)) {
-        track.browseRefined({
-          platform: "owa",
-          product: "teacher lesson resources",
-          engagementIntent: "refine",
-          componentType: "filter_link",
-          eventVersion: "2.0.0",
-          analyticsUseCase: "Teacher",
-          filterValue: "show results button",
-          filterType: "Subject filter",
-          activeFilters: {
-            content_types: "units",
-            learning_themes: theme,
-            categories: category,
-            year: year,
-          },
-        });
-      }
-      setIsMobileFilterDrawerOpen(false);
-      router.replace({
-        pathname: router.pathname,
-        query: params,
       });
     }
   };
